requirements.txt
langchain
streamlit
huggingface_hub
tiktoken
python-decouple
pypdf
pypdf2
sentence-transformers
torch
pytesseract
opencv-python
pdf2image
pandas
langchain-together
langchain_community
flask
pymongo
langchain-core
pdfplumber